Refactor db models
