+++ src/rust/plugin-registry/src/db/mod.rs
@@ -1,1 +1,2 @@
 pub mod client;
+pub mod models;
